c ur l

+++ src/main/java/pojlib/util/DownloadUtils.java
@@ -8,71 +8,42 @@ import pojlib.api.API_V1;
 import javax.net.ssl.SSLException;
 import java.io.*;
 import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.util.Objects;
 
 public class DownloadUtils {
+    public native static int cDownloadFile(String url, String filepath);
+    public native static String cResponseCodeString(int code);
 
-    private static void download(URL url, OutputStream os) throws IOException {
-        final int MAX_RETRIES = 3;
-        int attempts = 0;
-
-        while (true) {
-            try {
-                HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-                conn.setRequestProperty("User-Agent", "QuestCraft");
-                conn.setConnectTimeout(10000);
-                conn.setDoInput(true);
-                conn.connect();
-                if (conn.getResponseCode() == HttpURLConnection.HTTP_OK) {
-                    try (InputStream is = new StreamDL(conn.getInputStream())) {
-                        String[] segments = url.getPath().split("/");
-                        API_V1.currentDownload = segments[segments.length - 1];
-
-                        ((StreamDL)is).addListener((b, count) -> {
-
-                            if (b == -1) {
-                                API_V1.downloadStatus = 0;
-                                API_V1.currentDownload = null;
-                            } else {
-                                API_V1.downloadStatus = count * 0.000001;
-                            }
-
-                        });
-
-                        IOUtils.copy(is, os);
-                    }
-                    return;
-                }
-
-            } catch (IOException e) {
-                if (++attempts >= MAX_RETRIES || e instanceof SSLException) {
-                    throw new IOException("Unable to download from " + url, e);
-                }
-            }
-        }
-    }
-
-    public static void downloadFile(String url, File out) throws IOException {
+    public static void downloadFile(String url, File out) throws IOException, RuntimeException 
+    {
         Objects.requireNonNull(out.getParentFile()).mkdirs();
         File tempOut = File.createTempFile(out.getName(), ".part", out.getParentFile());
-        BufferedOutputStream bos = null;
-        try {
-            OutputStream bos2 = new BufferedOutputStream(Files.newOutputStream(tempOut.toPath()));
-            try {
-                download(new URL(url), bos2);
-                tempOut.renameTo(out);
-                bos2.close();
-                if (tempOut.exists()) tempOut.delete();
 
-            } catch (IOException th2) {
+        try 
+        {
+            int response = cDownloadFile(url, tempOut.getAbsolutePath());
+            if(response == 0) // CURLE_OK 
+            {
+                Logger.getInstance().appendToLog("Successfully downloaded a file from \"" + url + "\" to \"" + out.getName() + "\".");
+                tempOut.renameTo(out);
                 if (tempOut.exists()) tempOut.delete();
-                throw th2;
             }
-        } catch (IOException th3) {
+            else
+            {
+                if (tempOut.exists()) tempOut.delete();
+                String curlError = cResponseCodeString(response);
+                throw new IOException("Failed to download file from \"" + url + "\" to \"" + out.getName() + "\". cURL Error: " + curlError);
+            }
+        }
+        catch(Exception e)
+        {
             if (tempOut.exists()) tempOut.delete();
-            throw th3;
+            Logger.getInstance().appendToLog(e.toString());
+            throw e;
         }
     }
     public static boolean compareSHA1(File f, String sourceSHA) {
